+++ libcaf_core/caf/flow/observable.test.cpp
@@ -29,9 +29,9 @@ TEST("element_at(n) only takes the element with index n") {
       check_eq(collect(range(1, 3).element_at(0)), vector{1});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1)).element_at(0)), vector{1});
-      check_eq(collect(mat(range(1, 2)).element_at(0)), vector{1});
-      check_eq(collect(mat(range(1, 3)).element_at(0)), vector{1});
+      check_eq(collect(build(range(1, 1)).element_at(0)), vector{1});
+      check_eq(collect(build(range(1, 2)).element_at(0)), vector{1});
+      check_eq(collect(build(range(1, 3)).element_at(0)), vector{1});
     }
   }
   SECTION("element_at(1) picks the second element") {
@@ -40,8 +40,8 @@ TEST("element_at(n) only takes the element with index n") {
       check_eq(collect(range(1, 3).element_at(1)), vector{2});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 2)).element_at(1)), vector{2});
-      check_eq(collect(mat(range(1, 3)).element_at(1)), vector{2});
+      check_eq(collect(build(range(1, 2)).element_at(1)), vector{2});
+      check_eq(collect(build(range(1, 3)).element_at(1)), vector{2});
     }
   }
   SECTION("element_at(n) does not pick any element if n >= m") {
@@ -51,19 +51,17 @@ TEST("element_at(n) only takes the element with index n") {
       check_eq(collect(range(1, 3).element_at(3)), nil);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1)).element_at(1)), nil);
-      check_eq(collect(mat(range(1, 2)).element_at(2)), nil);
-      check_eq(collect(mat(range(1, 3)).element_at(3)), nil);
+      check_eq(collect(build(range(1, 1)).element_at(1)), nil);
+      check_eq(collect(build(range(1, 2)).element_at(2)), nil);
+      check_eq(collect(build(range(1, 3)).element_at(3)), nil);
     }
   }
   SECTION("element_at(n) propagates errors") {
     SECTION("blueprint") {
-      check_eq(collect(obs_error(sec::runtime_error).element_at(1)),
-               sec::runtime_error);
+      check_eq(collect(obs_error().element_at(1)), sec::runtime_error);
     }
     SECTION("observable") {
-      check_eq(collect(mat(obs_error(sec::runtime_error)).element_at(1)),
-               sec::runtime_error);
+      check_eq(collect(build(obs_error()).element_at(1)), sec::runtime_error);
     }
   }
 }
@@ -77,9 +75,9 @@ TEST("reduce(init, fn) combines all items into one final value using fn") {
       check_eq(collect(range(1, 3).reduce(0, adder)), vector{6});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1)).reduce(0, adder)), vector{1});
-      check_eq(collect(mat(range(1, 2)).reduce(0, adder)), vector{3});
-      check_eq(collect(mat(range(1, 3)).reduce(0, adder)), vector{6});
+      check_eq(collect(build(range(1, 1)).reduce(0, adder)), vector{1});
+      check_eq(collect(build(range(1, 2)).reduce(0, adder)), vector{3});
+      check_eq(collect(build(range(1, 3)).reduce(0, adder)), vector{6});
     }
   }
   SECTION("reduce(init, fn) emits 0 for a range of size 0") {
@@ -87,7 +85,7 @@ TEST("reduce(init, fn) combines all items into one final value using fn") {
       check_eq(collect(range(1, 0).reduce(0, adder)), vector{0});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 0)).reduce(0, adder)), vector{0});
+      check_eq(collect(build(range(1, 0)).reduce(0, adder)), vector{0});
     }
   }
   SECTION("reduce(init, fn) propagates errors") {
@@ -99,11 +97,12 @@ TEST("reduce(init, fn) combines all items into one final value using fn") {
       check_eq(collect(obs_error().reduce(0, adder)), sec::runtime_error);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1).concat(obs_error())).reduce(0, adder)),
+      check_eq(collect(build(range(1, 1).concat(obs_error())).reduce(0, adder)),
                sec::runtime_error);
-      check_eq(collect(mat(range(1, 3).concat(obs_error())).reduce(0, adder)),
+      check_eq(collect(build(range(1, 3).concat(obs_error())).reduce(0, adder)),
                sec::runtime_error);
-      check_eq(collect(mat(obs_error()).reduce(0, adder)), sec::runtime_error);
+      check_eq(collect(build(obs_error()).reduce(0, adder)),
+               sec::runtime_error);
     }
   }
 }
@@ -117,9 +116,9 @@ TEST("scan(init, fn) creates successive reduced values using fn") {
       check_eq(collect(range(1, 3).scan(0, adder)), vector{1, 3, 6});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1)).scan(0, adder)), vector{1});
-      check_eq(collect(mat(range(1, 2)).scan(0, adder)), vector{1, 3});
-      check_eq(collect(mat(range(1, 3)).scan(0, adder)), vector{1, 3, 6});
+      check_eq(collect(build(range(1, 1)).scan(0, adder)), vector{1});
+      check_eq(collect(build(range(1, 2)).scan(0, adder)), vector{1, 3});
+      check_eq(collect(build(range(1, 3)).scan(0, adder)), vector{1, 3, 6});
     }
   }
   SECTION("scan(init, fn) emits 0 for a range of size 0") {
@@ -127,7 +126,7 @@ TEST("scan(init, fn) creates successive reduced values using fn") {
       check_eq(collect(range(1, 0).scan(0, adder)), nil);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 0)).scan(0, adder)), nil);
+      check_eq(collect(build(range(1, 0)).scan(0, adder)), nil);
     }
   }
   SECTION("scan(init, fn) propagates errors") {
@@ -139,11 +138,11 @@ TEST("scan(init, fn) creates successive reduced values using fn") {
       check_eq(collect(obs_error().scan(0, adder)), sec::runtime_error);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1).concat(obs_error())).scan(0, adder)),
+      check_eq(collect(build(range(1, 1).concat(obs_error())).scan(0, adder)),
                sec::runtime_error);
-      check_eq(collect(mat(range(1, 3).concat(obs_error())).scan(0, adder)),
+      check_eq(collect(build(range(1, 3).concat(obs_error())).scan(0, adder)),
                sec::runtime_error);
-      check_eq(collect(mat(obs_error()).scan(0, adder)), sec::runtime_error);
+      check_eq(collect(build(obs_error()).scan(0, adder)), sec::runtime_error);
     }
   }
 }
@@ -157,10 +156,10 @@ TEST("skip(n) skips the first n elements in a range of size m") {
       check_eq(collect(range(1, 3).skip(0)), vector{1, 2, 3});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 0)).skip(0)), nil);
-      check_eq(collect(mat(range(1, 1)).skip(0)), vector{1});
-      check_eq(collect(mat(range(1, 2)).skip(0)), vector{1, 2});
-      check_eq(collect(mat(range(1, 3)).skip(0)), vector{1, 2, 3});
+      check_eq(collect(build(range(1, 0)).skip(0)), nil);
+      check_eq(collect(build(range(1, 1)).skip(0)), vector{1});
+      check_eq(collect(build(range(1, 2)).skip(0)), vector{1, 2});
+      check_eq(collect(build(range(1, 3)).skip(0)), vector{1, 2, 3});
     }
   }
   SECTION("skip(n) truncates the first n elements if n < m") {
@@ -169,8 +168,8 @@ TEST("skip(n) skips the first n elements in a range of size m") {
       check_eq(collect(range(1, 3).skip(2)), vector{3});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 3)).skip(1)), vector{2, 3});
-      check_eq(collect(mat(range(1, 3)).skip(2)), vector{3});
+      check_eq(collect(build(range(1, 3)).skip(1)), vector{2, 3});
+      check_eq(collect(build(range(1, 3)).skip(2)), vector{3});
     }
   }
   SECTION("skip(n) drops all elements if n >= m") {
@@ -179,8 +178,8 @@ TEST("skip(n) skips the first n elements in a range of size m") {
       check_eq(collect(range(1, 3).skip(4)), nil);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 3)).skip(3)), nil);
-      check_eq(collect(mat(range(1, 3)).skip(4)), nil);
+      check_eq(collect(build(range(1, 3)).skip(3)), nil);
+      check_eq(collect(build(range(1, 3)).skip(4)), nil);
     }
   }
   SECTION("skip(n) stops early when the next operator stops early") {
@@ -192,11 +191,11 @@ TEST("skip(n) skips the first n elements in a range of size m") {
       check_eq(collect(range(1, 5).skip(2).take(4)), vector{3, 4, 5});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 5)).skip(2).take(0)), nil);
-      check_eq(collect(mat(range(1, 5)).skip(2).take(1)), vector{3});
-      check_eq(collect(mat(range(1, 5)).skip(2).take(2)), vector{3, 4});
-      check_eq(collect(mat(range(1, 5)).skip(2).take(3)), vector{3, 4, 5});
-      check_eq(collect(mat(range(1, 5)).skip(2).take(4)), vector{3, 4, 5});
+      check_eq(collect(build(range(1, 5)).skip(2).take(0)), nil);
+      check_eq(collect(build(range(1, 5)).skip(2).take(1)), vector{3});
+      check_eq(collect(build(range(1, 5)).skip(2).take(2)), vector{3, 4});
+      check_eq(collect(build(range(1, 5)).skip(2).take(3)), vector{3, 4, 5});
+      check_eq(collect(build(range(1, 5)).skip(2).take(4)), vector{3, 4, 5});
     }
   }
   SECTION("skip(n) forwards errors") {
@@ -205,8 +204,8 @@ TEST("skip(n) skips the first n elements in a range of size m") {
       check_eq(collect(obs_error().skip(1)), sec::runtime_error);
     }
     SECTION("observable") {
-      check_eq(collect(mat(obs_error()).skip(0)), sec::runtime_error);
-      check_eq(collect(mat(obs_error()).skip(1)), sec::runtime_error);
+      check_eq(collect(build(obs_error()).skip(0)), sec::runtime_error);
+      check_eq(collect(build(obs_error()).skip(1)), sec::runtime_error);
     }
   }
 }
@@ -219,9 +218,9 @@ TEST("first() returns the first element in a range of size m") {
       check_eq(collect(range(1, 3).first()), vector{1});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1)).first()), vector{1});
-      check_eq(collect(mat(range(1, 2)).first()), vector{1});
-      check_eq(collect(mat(range(1, 3)).first()), vector{1});
+      check_eq(collect(build(range(1, 1)).first()), vector{1});
+      check_eq(collect(build(range(1, 2)).first()), vector{1});
+      check_eq(collect(build(range(1, 3)).first()), vector{1});
     }
   }
   SECTION("first() returns an empty range if the input is empty") {
@@ -229,7 +228,7 @@ TEST("first() returns the first element in a range of size m") {
       check_eq(collect(range(1, 0).first()), nil);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 0)).first()), nil);
+      check_eq(collect(build(range(1, 0)).first()), nil);
     }
   }
 }
@@ -242,9 +241,9 @@ TEST("last() returns the last element in a range of size m") {
       check_eq(collect(range(1, 3).last()), vector{3});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1)).last()), vector{1});
-      check_eq(collect(mat(range(1, 2)).last()), vector{2});
-      check_eq(collect(mat(range(1, 3)).last()), vector{3});
+      check_eq(collect(build(range(1, 1)).last()), vector{1});
+      check_eq(collect(build(range(1, 2)).last()), vector{2});
+      check_eq(collect(build(range(1, 3)).last()), vector{3});
     }
   }
   SECTION("last() returns an empty range if the input is empty") {
@@ -252,7 +251,7 @@ TEST("last() returns the last element in a range of size m") {
       check_eq(collect(range(1, 0).last()), nil);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 0)).last()), nil);
+      check_eq(collect(build(range(1, 0)).last()), nil);
     }
   }
 }
@@ -266,65 +265,57 @@ TEST("on_error_complete() suppresses errors") {
       check_eq(collect(range(1, 0).on_error_complete()), nil);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1)).on_error_complete()), vector{1});
-      check_eq(collect(mat(range(1, 2)).on_error_complete()), vector{1, 2});
-      check_eq(collect(mat(range(1, 3)).on_error_complete()), vector{1, 2, 3});
-      check_eq(collect(mat(range(1, 0)).on_error_complete()), nil);
+      check_eq(collect(build(range(1, 1)).on_error_complete()), vector{1});
+      check_eq(collect(build(range(1, 2)).on_error_complete()), vector{1, 2});
+      check_eq(collect(build(range(1, 3)).on_error_complete()),
+               vector{1, 2, 3});
+      check_eq(collect(build(range(1, 0)).on_error_complete()), nil);
     }
   }
   SECTION("on_error_complete() does not propagate errors") {
     SECTION("blueprint") {
-      check_eq(collect(range(1, 1)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_complete()),
+      check_eq(collect(range(1, 1).concat(obs_error()).on_error_complete()),
                vector{1});
-      check_eq(collect(range(1, 2)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_complete()),
+      check_eq(collect(range(1, 2).concat(obs_error()).on_error_complete()),
                vector{1, 2});
-      check_eq(collect(range(1, 3)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_complete()),
+      check_eq(collect(range(1, 3).concat(obs_error()).on_error_complete()),
                vector{1, 2, 3});
-      check_eq(collect(obs_error(sec::runtime_error).on_error_complete()), nil);
+      check_eq(collect(obs_error().on_error_complete()), nil);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1).concat(obs_error(sec::runtime_error)))
-                         .on_error_complete()),
+      check_eq(collect(
+                 build(range(1, 1).concat(obs_error())).on_error_complete()),
                vector{1});
-      check_eq(collect(mat(range(1, 2).concat(obs_error(sec::runtime_error)))
-                         .on_error_complete()),
+      check_eq(collect(
+                 build(range(1, 2).concat(obs_error())).on_error_complete()),
                vector{1, 2});
-      check_eq(collect(mat(range(1, 3).concat(obs_error(sec::runtime_error)))
-                         .on_error_complete()),
+      check_eq(collect(
+                 build(range(1, 3).concat(obs_error())).on_error_complete()),
                vector{1, 2, 3});
-      check_eq(collect(mat(obs_error(sec::runtime_error)).on_error_complete()),
-               nil);
+      check_eq(collect(build(obs_error()).on_error_complete()), nil);
     }
   }
   SECTION("on_error_complete() completes on error") {
     SECTION("blueprint") {
       check_eq(collect(range(1, 2)
-                         .concat(obs_error(sec::runtime_error))
+                         .concat(obs_error())
                          .concat(range(1, 2))
                          .on_error_complete()),
                vector{1, 2});
       check_eq(collect(range(1, 3)
-                         .concat(obs_error(sec::runtime_error))
+                         .concat(obs_error())
                          .concat(range(1, 3))
                          .on_error_complete()),
                vector{1, 2, 3});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 2)
-                             .concat(obs_error(sec::runtime_error))
-                             .concat(range(1, 2)))
-                         .on_error_complete()),
+      check_eq(collect(
+                 build(range(1, 2).concat(obs_error()).concat(range(1, 2)))
+                   .on_error_complete()),
                vector{1, 2});
-      check_eq(collect(mat(range(1, 3)
-                             .concat(obs_error(sec::runtime_error))
-                             .concat(range(1, 3)))
-                         .on_error_complete()),
+      check_eq(collect(
+                 build(range(1, 3).concat(obs_error()).concat(range(1, 3)))
+                   .on_error_complete()),
                vector{1, 2, 3});
     }
   }
@@ -339,60 +330,52 @@ TEST("on_error_return_item() replaces an error with a value") {
       check_eq(collect(range(1, 0).on_error_return_item(42)), nil);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1)).on_error_return_item(42)), vector{1});
-      check_eq(collect(mat(range(1, 2)).on_error_return_item(42)),
+      check_eq(collect(build(range(1, 1)).on_error_return_item(42)), vector{1});
+      check_eq(collect(build(range(1, 2)).on_error_return_item(42)),
                vector{1, 2});
-      check_eq(collect(mat(range(1, 3)).on_error_return_item(42)),
+      check_eq(collect(build(range(1, 3)).on_error_return_item(42)),
                vector{1, 2, 3});
-      check_eq(collect(mat(range(1, 0)).on_error_return_item(42)), nil);
+      check_eq(collect(build(range(1, 0)).on_error_return_item(42)), nil);
     }
   }
   SECTION("on_error_return_item() returns the item when an error occurs") {
     SECTION("blueprint") {
-      check_eq(collect(range(1, 0)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_return_item(42)),
-               vector{42});
+      check_eq(collect(obs_error().on_error_return_item(42)), vector{42});
       check_eq(collect(range(1, 2)
-                         .concat(obs_error(sec::runtime_error))
+                         .concat(obs_error())
                          .concat(range(1, 2))
                          .on_error_return_item(42)),
                vector{1, 2, 42});
       check_eq(collect(range(1, 3)
-                         .concat(obs_error(sec::runtime_error))
+                         .concat(obs_error())
                          .concat(range(1, 3))
                          .on_error_return_item(42)),
                vector{1, 2, 3, 42});
-      check_eq(collect(range(1, 2)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_return_item(42)
-                         .take(2)),
-               vector{1, 2});
-      check_eq(collect(range(1, 3)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_return_item(42)
-                         .take(3)),
-               vector{1, 2, 3});
+      check_eq(
+        collect(
+          range(1, 2).concat(obs_error()).on_error_return_item(42).take(2)),
+        vector{1, 2});
+      check_eq(
+        collect(
+          range(1, 3).concat(obs_error()).on_error_return_item(42).take(3)),
+        vector{1, 2, 3});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 0).concat(obs_error(sec::runtime_error)))
-                         .on_error_return_item(42)),
+      check_eq(collect(build(obs_error()).on_error_return_item(42)),
                vector{42});
-      check_eq(collect(mat(range(1, 2)
-                             .concat(obs_error(sec::runtime_error))
-                             .concat(range(1, 2)))
-                         .on_error_return_item(42)),
+      check_eq(collect(
+                 build(range(1, 2).concat(obs_error()).concat(range(1, 2)))
+                   .on_error_return_item(42)),
                vector{1, 2, 42});
-      check_eq(collect(mat(range(1, 3)
-                             .concat(obs_error(sec::runtime_error))
-                             .concat(range(1, 3)))
-                         .on_error_return_item(42)),
+      check_eq(collect(
+                 build(range(1, 3).concat(obs_error()).concat(range(1, 3)))
+                   .on_error_return_item(42)),
                vector{1, 2, 3, 42});
-      check_eq(collect(mat(range(1, 2).concat(obs_error(sec::runtime_error)))
+      check_eq(collect(build(range(1, 2).concat(obs_error()))
                          .on_error_return_item(42)
                          .take(2)),
                vector{1, 2});
-      check_eq(collect(mat(range(1, 3).concat(obs_error(sec::runtime_error)))
+      check_eq(collect(build(range(1, 3).concat(obs_error()))
                          .on_error_return_item(42)
                          .take(3)),
                vector{1, 2, 3});
@@ -417,19 +400,19 @@ TEST("on_error_return() replaces an error with a value") {
                nil);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 1)).on_error_return([](const error&) {
+      check_eq(collect(build(range(1, 1)).on_error_return([](const error&) {
                  return expected<int>{42};
                })),
                vector{1});
-      check_eq(collect(mat(range(1, 2)).on_error_return([](const error&) {
+      check_eq(collect(build(range(1, 2)).on_error_return([](const error&) {
                  return expected<int>{42};
                })),
                vector{1, 2});
-      check_eq(collect(mat(range(1, 3)).on_error_return([](const error&) {
+      check_eq(collect(build(range(1, 3)).on_error_return([](const error&) {
                  return expected<int>{42};
                })),
                vector{1, 2, 3});
-      check_eq(collect(mat(range(1, 0)).on_error_return([](const error&) {
+      check_eq(collect(build(range(1, 0)).on_error_return([](const error&) {
                  return expected<int>{42};
                })),
                nil);
@@ -438,50 +421,46 @@ TEST("on_error_return() replaces an error with a value") {
   SECTION("on_error_return() replaces an error with a value from the handler") {
     auto return_42 = [](const error&) { return expected<int>{42}; };
     SECTION("blueprint") {
-      check_eq(collect(range(1, 0)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_return(return_42)),
+      check_eq(collect(
+                 range(1, 0).concat(obs_error()).on_error_return(return_42)),
                vector{42});
       check_eq(collect(range(1, 2)
-                         .concat(obs_error(sec::runtime_error))
+                         .concat(obs_error())
                          .concat(range(1, 2))
                          .on_error_return(return_42)),
                vector{1, 2, 42});
       check_eq(collect(range(1, 3)
-                         .concat(obs_error(sec::runtime_error))
+                         .concat(obs_error())
                          .concat(range(1, 3))
                          .on_error_return(return_42)),
                vector{1, 2, 3, 42});
-      check_eq(collect(range(1, 2)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_return(return_42)
-                         .take(2)),
-               vector{1, 2});
-      check_eq(collect(range(1, 3)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_return(return_42)
-                         .take(3)),
-               vector{1, 2, 3});
+      check_eq(
+        collect(
+          range(1, 2).concat(obs_error()).on_error_return(return_42).take(2)),
+        vector{1, 2});
+      check_eq(
+        collect(
+          range(1, 3).concat(obs_error()).on_error_return(return_42).take(3)),
+        vector{1, 2, 3});
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 0).concat(obs_error(sec::runtime_error)))
-                         .on_error_return(return_42)),
-               vector{42});
-      check_eq(collect(mat(range(1, 2)
-                             .concat(obs_error(sec::runtime_error))
-                             .concat(range(1, 2)))
-                         .on_error_return(return_42)),
+      check_eq(
+        collect(
+          build(range(1, 0).concat(obs_error())).on_error_return(return_42)),
+        vector{42});
+      check_eq(collect(
+                 build(range(1, 2).concat(obs_error()).concat(range(1, 2)))
+                   .on_error_return(return_42)),
                vector{1, 2, 42});
-      check_eq(collect(mat(range(1, 3)
-                             .concat(obs_error(sec::runtime_error))
-                             .concat(range(1, 3)))
-                         .on_error_return(return_42)),
+      check_eq(collect(
+                 build(range(1, 3).concat(obs_error()).concat(range(1, 3)))
+                   .on_error_return(return_42)),
                vector{1, 2, 3, 42});
-      check_eq(collect(mat(range(1, 2).concat(obs_error(sec::runtime_error)))
+      check_eq(collect(build(range(1, 2).concat(obs_error()))
                          .on_error_return(return_42)
                          .take(2)),
                vector{1, 2});
-      check_eq(collect(mat(range(1, 3).concat(obs_error(sec::runtime_error)))
+      check_eq(collect(build(range(1, 3).concat(obs_error()))
                          .on_error_return(return_42)
                          .take(3)),
                vector{1, 2, 3});
@@ -493,47 +472,42 @@ TEST("on_error_return() replaces an error with a value") {
     };
     auto return_err = [](const error& err) { return expected<int>{err}; };
     SECTION("blueprint") {
-      check_eq(collect(range(1, 0)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_return(return_unexpected)),
+      check_eq(collect(obs_error().on_error_return(return_unexpected)),
                sec::unexpected_message);
       check_eq(collect(range(1, 2)
-                         .concat(obs_error(sec::runtime_error))
+                         .concat(obs_error())
                          .concat(range(1, 2))
                          .on_error_return(return_unexpected)),
                sec::unexpected_message);
       check_eq(collect(range(1, 3)
-                         .concat(obs_error(sec::runtime_error))
+                         .concat(obs_error())
                          .concat(range(1, 3))
                          .on_error_return(return_err)),
                sec::runtime_error);
-      check_eq(collect(range(1, 2)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_return(return_err)),
+      check_eq(collect(
+                 range(1, 2).concat(obs_error()).on_error_return(return_err)),
                sec::runtime_error);
-      check_eq(collect(range(1, 3)
-                         .concat(obs_error(sec::runtime_error))
-                         .on_error_return(return_unexpected)),
-               sec::unexpected_message);
+      check_eq(
+        collect(
+          range(1, 3).concat(obs_error()).on_error_return(return_unexpected)),
+        sec::unexpected_message);
     }
     SECTION("observable") {
-      check_eq(collect(mat(range(1, 0).concat(obs_error(sec::runtime_error)))
-                         .on_error_return(return_unexpected)),
+      check_eq(collect(build(obs_error()).on_error_return(return_unexpected)),
                sec::unexpected_message);
-      check_eq(collect(mat(range(1, 2)
-                             .concat(obs_error(sec::runtime_error))
-                             .concat(range(1, 2)))
-                         .on_error_return(return_unexpected)),
+      check_eq(collect(
+                 build(range(1, 2).concat(obs_error()).concat(range(1, 2)))
+                   .on_error_return(return_unexpected)),
                sec::unexpected_message);
-      check_eq(collect(mat(range(1, 3)
-                             .concat(obs_error(sec::runtime_error))
-                             .concat(range(1, 3)))
-                         .on_error_return(return_err)),
+      check_eq(collect(
+                 build(range(1, 3).concat(obs_error()).concat(range(1, 3)))
+                   .on_error_return(return_err)),
                sec::runtime_error);
-      check_eq(collect(mat(range(1, 2).concat(obs_error(sec::runtime_error)))
-                         .on_error_return(return_err)),
-               sec::runtime_error);
-      check_eq(collect(mat(range(1, 3).concat(obs_error(sec::runtime_error)))
+      check_eq(
+        collect(
+          build(range(1, 2).concat(obs_error())).on_error_return(return_err)),
+        sec::runtime_error);
+      check_eq(collect(build(range(1, 3).concat(obs_error()))
                          .on_error_return(return_unexpected)),
                sec::unexpected_message);
     }